+++ app/changelog.md
@@ -355,4 +355,5 @@
 - Added getMenu function to triMenu that will grab any menu item using its id attribute
 - Changed Todo example to use new triMenu to change it's badge value based on number of todos
 - Added click handler to menus to allow click events as well as state router changes
+- Added openInNewTab option to menu items
 - Updated Angular Permissions to 3.2.1